requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
gspread==6.1.0
google-auth==2.35.0
python-dotenv==1.0.1
feedparser==6.0.11
openai==1.51.0
beautifulsoup4==4.12.3
httpx==0.27.2
supabase==2.15.2
